Move deconz event tests to test_deconz_event.py
